chore: restructure sc filters, requests and account tabs
